lab 8

+++ symbols/src/main/java/com/example/symbols/repository/CountResultRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CountResultRepository extends JpaRepository<CountResultEntity, Integer> {
+public interface CountResultRepository extends JpaRepository<CountResultEntity, Long> {
 }
